Cast the variable to compile also for MuJoCo 3.5.0

+++ mujoco_ros2_control/src/mujoco_system_interface.cpp
@@ -459,8 +459,8 @@ mjModel* loadModelFromTopic(rclcpp::Node::SharedPtr node, const std::string& top
       mj_deleteSpec(spec);
     }
     mj_deleteSpec(spec);
-    RCLCPP_INFO(node->get_logger(), "Model body count: %d", mnew->nbody);
-    RCLCPP_INFO(node->get_logger(), "Model geom count: %d", mnew->ngeom);
+    RCLCPP_INFO(node->get_logger(), "Model body count: %ld", static_cast<long>(mnew->nbody));
+    RCLCPP_INFO(node->get_logger(), "Model geom count: %ld", static_cast<long>(mnew->ngeom));
   }
   return mnew;
 }
@@ -1631,7 +1631,8 @@ bool MujocoSystemInterface::register_mujoco_actuators()
 
   for (int i = 0; i < mj_model_->nu; i++)
   {
-    RCLCPP_DEBUG(get_logger(), "Registering MuJoCo actuator %d/%d", i + 1, mj_model_->nu);
+    RCLCPP_DEBUG(get_logger(), "Registering MuJoCo actuator %ld/%ld", static_cast<long>(i + 1),
+                 static_cast<long>(mj_model_->nu));
     MuJoCoActuatorData& actuator_data = mujoco_actuator_data_.at(i);
 
     // Get the name of the joint/tendon corresponding to the actuator ID
@@ -2706,8 +2707,9 @@ void MujocoSystemInterface::PhysicsLoop()
             sim_->speed_changed = false;
 
             // Copy data to the control
-            mju_copy(mj_data_->ctrl, mj_data_control_->ctrl, mj_model_->nu);
-            mju_copy(mj_data_->qfrc_applied, mj_data_control_->qfrc_applied, mj_model_->nu);
+            mju_copy(mj_data_->ctrl, mj_data_control_->ctrl, static_cast<int>(mj_model_->nu));
+            mju_copy(mj_data_->qfrc_applied, mj_data_control_->qfrc_applied,
+                     static_cast<int>(mj_model_->nu));
             // run single step, let next iteration deal with timing
             mj_step(mj_model_, mj_data_);
 
@@ -2757,8 +2759,9 @@ void MujocoSystemInterface::PhysicsLoop()
 #endif
 
               // Copy data to the control
-              mju_copy(mj_data_->ctrl, mj_data_control_->ctrl, mj_model_->nu);
-              mju_copy(mj_data_->qfrc_applied, mj_data_control_->qfrc_applied, mj_model_->nu);
+               mju_copy(mj_data_->ctrl, mj_data_control_->ctrl, static_cast<int>(mj_model_->nu));
+               mju_copy(mj_data_->qfrc_applied, mj_data_control_->qfrc_applied,
+                        static_cast<int>(mj_model_->nu));
               // call mj_step
               mj_step(mj_model_, mj_data_);
